averager3/requirements.txt
streamlit
matplotlib
pandas
pyyaml
plotly
